test: add test units for login and recovery forms
